backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
SpeechRecognition
websockets
flask
flask-cors
google-generativeai
sounddevice
numpy
python-dotenv
SpeechRecognition